Add files via upload
This is the TIP CALCULATOR V2 program
Assuming that you and your friends went out for lunch. The program calculates the amount of money that will be paid by each of you if you were to equally share the bill and tip.

It is an upgraded version of the TIP CALCULATOR I created earlier.